MR-96: add the form
