+++ src/Components/Layout/index.js
@@ -25,9 +25,10 @@ function Layout() {
             <Route path="/classes" component={Classes} />
             <Route path="/members" component={Members} />
             <Route path="/subscriptions" component={Subscriptions} />
-            <Route exact path="/superAdmins" component={SuperAdmins} />
+            <Route path="/superAdmins" exact component={SuperAdmins} />
             <Route path="/trainers" component={Trainers} />
-            <Route path="/superAdmins/form" component={FormSuperAdmin} />
+            <Route path="/superAdmins/form" exact component={FormSuperAdmin} />
+            <Route path="/superAdmins/form/:id" component={FormSuperAdmin} />
           </Switch>
         </div>
         <Footer />
